# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Filkx API",
        "description": "API documentation for the Tech Filkx e-commerce platform",
        "contact": {
            "email": "admin@filkx.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://dev.api.tech.filkx.com",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get the current cart (guest session or authenticated user)",
                "operationId": "11d5201ddac04df2c82b6741daba911b",
                "parameters": [
                    {
                        "name": "X-Cart-Session-ID",
                        "in": "header",
                        "description": "Guest cart session identifier. Ignored for authenticated requests, where the cart is resolved from the bearer token instead.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a product variant to the cart",
                "operationId": "8bdc675b00982b482e9edae2dddf34d8",
                "parameters": [
                    {
                        "name": "X-Cart-Session-ID",
                        "in": "header",
                        "description": "Guest cart session identifier. Ignored for authenticated requests, where the cart is resolved from the bearer token instead.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "variantId"
                                ],
                                "properties": {
                                    "variantId": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1,
                                        "minimum": 1
                                    },
                                    "sessionId": {
                                        "description": "Alternative to the X-Cart-Session-ID header",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated cart contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product or variant not found"
                    },
                    "422": {
                        "description": "Requested quantity exceeds available stock, or validation failed"
                    }
                }
            }
        },
        "/api/v1/cart/items/{itemId}": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "Update the quantity of a cart item",
                "operationId": "f2cb2af5baab1d4cd2e9e6cd3496edd9",
                "parameters": [
                    {
                        "name": "X-Cart-Session-ID",
                        "in": "header",
                        "description": "Guest cart session identifier. Ignored for authenticated requests, where the cart is resolved from the bearer token instead.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated cart contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item does not exist or does not belong to this cart/session"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove an item from the cart",
                "operationId": "ea9748540900b0183277283a636576a7",
                "parameters": [
                    {
                        "name": "X-Cart-Session-ID",
                        "in": "header",
                        "description": "Guest cart session identifier. Ignored for authenticated requests, where the cart is resolved from the bearer token instead.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated cart contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cart/merge": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Merge a guest cart into the authenticated user's cart",
                "operationId": "6682305248a20bfc6eb47626fa98eb06",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sessionId"
                                ],
                                "properties": {
                                    "sessionId": {
                                        "description": "Guest cart session id to merge into the authenticated cart",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Merged cart contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CartItemResource": {
                "title": "Cart Item Resource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "variantId": {
                        "type": "integer"
                    },
                    "productId": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "oldPrice": {
                        "type": "number",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "stock": {
                        "description": "Available stock (quantity minus reserved), clamped to the item quantity",
                        "type": "integer"
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "subtotal": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "CartResource": {
                "title": "Cart Resource",
                "properties": {
                    "sessionId": {
                        "description": "Guest cart session identifier; null for an authenticated user's cart",
                        "type": "string",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItemResource"
                        }
                    },
                    "total": {
                        "type": "number"
                    },
                    "promotionDiscount": {
                        "type": "number"
                    },
                    "discountedTotal": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}